+++ packages/docs/astro.config.ts
@@ -65,20 +65,20 @@ const sidebar = [
   },
   chatkitTypeDocSidebarGroup,
   {
-    label: 'Managed backend',
+    label: 'Self-hosted backend',
     items: [
       {
-        label: 'ChatKit API',
-        link: 'https://platform.openai.com/docs/guides/chatkit',
+        label: 'ChatKit Python SDK',
+        link: 'https://openai.github.io/chatkit-python',
       },
     ],
   },
   {
-    label: 'Self-hosted backend',
+    label: 'OpenAI-hosted backend',
     items: [
       {
-        label: 'Python SDK',
-        link: 'https://openai.github.io/chatkit-python',
+        label: 'Managed ChatKit Docs',
+        link: 'https://platform.openai.com/docs/guides/chatkit',
       },
     ],
   },
@@ -119,6 +119,13 @@ export default defineConfig({
         Sidebar: './src/components/Sidebar.astro',
         MobileMenuFooter: './src/components/MobileFooter.astro',
       },
+      social: [
+        {
+          icon: 'github',
+          label: 'ChatKit.js',
+          href: 'https://github.com/openai/chatkit-js',
+        },
+      ],
       locales: {
         root: {
           label: 'English',